// tests/types.ts
export type TestAxisOptions = {
  request: APIRequestContext;
  expectedStatus: number;
  expectValidity: boolean;
  minQuality?: number | string;
  expectedError?: string;
};

export type Layer = {
  id: string;
};

export type AxisData = {
  label: null;
  steps: { label: null; value: number }[];
  datasetStats: {
    minValue: number;
    maxValue: number;
    mean: number;
    stddev: number;
  };
  quotient: [string, string];
  quotients: {
    name: string;
    label: string;
    emoji?: string;
    maxZoom: number;
    description: string;
    copyrights: string[];
    direction: [string][];
    unit: {
      id: string;
      shortName: string;
      longName: string;
    };
  }[];
  quality: number;
  transformation: {
    transformation: string;
    mean: number;
    skew: number;
    stddev: number;
    lowerBound: number;
    upperBound: number;
  };
  parent: null;
}[];